delete reference_kv.py file
